Added email, user page and DI

+++ app/Http/Controllers/BlackListedFileController.php
@@ -15,12 +15,11 @@ class BlackListedFileController extends Controller
     public function blacklist(Request $request)
     {
         if (!Auth::user() || !(Auth::user()->role == "moderator" || Auth::user()->role == "admin")) {
-            return response()->json(['message' => 'Forbidden'], 403);
+            return view('error',['message'=>'You don\'t have the rights to be here.']);
         }
         if (($request->file('files')) == null) {
-            return response()->json([
-                'message' => 'You must upload at least one file.'
-            ], 422);
+            return view('error',['message'=>'You must upload at least one file.']);
+
         }
         foreach ($request->file('files') as $file) {
             $blacklistedFile = new BlacklistedFile();
@@ -34,11 +33,17 @@ class BlackListedFileController extends Controller
         ], 201);
     }
     public function getBlackListedFiles(){
+        if (!Auth::user() || !(Auth::user()->role == "moderator" || Auth::user()->role == "admin")) {
+            return view('error',['message'=>'You don\'t have the rights to be here.']);
+        }
         $files = $this->blacklistedfilesRepository->getBlacklistedfiles();
         return view('tables/listBlacklistedFiles', ['files' => $files]);
     }
 
     public function deleteBlacklistedFile($id){
+        if (!Auth::user() || !(Auth::user()->role == "moderator" || Auth::user()->role == "admin")) {
+            return view('error',['message'=>'You don\'t have the rights to be here.']);
+        }
         $file = $this->blacklistedfilesRepository->getFileById($id);
         $this->blacklistedfilesRepository->deleteFile($file);
         return view ('tables/listBlacklistedFiles');
